requirements.txt
Django==3.2.2
dj-database-url==0.5.0
psycopg2-binary==2.9.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
#pytest==8.2.2
#pytest-django==4.8.0